исправил замечания наставника
