+++ js/main.js
@@ -1,29 +1,2 @@
-import './data.js';
-import './util.js';
-
-const MESSAGE = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-const NAME = [
-  'Максим',
-  'Сергей',
-  'Игорь',
-  'Виталий',
-  'Петр',
-  'Михаил'
-];
-const DESCRIPTION = [
-  'Фото из автобуса',
-  'Я на море',
-  'Вместе идем на работу',
-  'Завтра в школу',
-  'Выходные в лесу',
-  'На рыбалке'
-];
-
-export {MESSAGE, NAME, DESCRIPTION};
+import {photosMock} from './data.js';
+photosMock();
